Merge parsers and commands into single files
I think having these separated makes it a little harder to work on each command. In this commit, I've moved the logic from each parser file to a new method named 'define_args' in the corresponding command file. I had previously added the name and description to the command file, so now we can change the map_of_commands into a simple list of all command files to act on.

Also updated tests so that the setup step refers to the new define_args method.