Automate finding missing strings
There is a group of strings I appear to have added manually that was deleted with the new loc file dump, and another group of string keys that are used in code and don't appear in properties.
Adding here to not lose them, but hope to find they are not necessary.